[libromdata] DMGPrivate::RomSize(): Reduce arrays from int[] to uint16_t[].
Code size difference: (64-bit Gentoo Linux, gcc-13.2.0, release build, no LTO)

   text    data     bss     dec     hex filename
  26361    1216       0   27577    6bb9 DMG.cpp.o [before]
  26323    1216       0   27539    6b93 DMG.cpp.o [after]
    -38       0       0     -38     -26 Difference

Interestingly, un-inling the function *increased* code size:

   text    data     bss     dec     hex filename
  26361    1216       0   27577    6bb9 DMG.cpp.o [before]
  26477    1216       0   27693    6c2d DMG.cpp.o [un-inlined]
  26429    1216       0   27645    6bfd DMG.cpp.o [un-inlined, uint16_t]

+++ src/libromdata/Handheld/DMG.cpp
@@ -377,8 +377,8 @@ inline DMGPrivate::dmg_cart_type DMGPrivate::CartType(uint8_t type)
  */
 inline int DMGPrivate::RomSize(uint8_t type)
 {
-	static const int rom_size[] = {32, 64, 128, 256, 512, 1024, 2048, 4096};
-	static const int rom_size_52[] = {1152, 1280, 1536};
+	static const uint16_t rom_size[] = {32, 64, 128, 256, 512, 1024, 2048, 4096};
+	static const uint16_t rom_size_52[] = {1152, 1280, 1536};
 	if (type < ARRAY_SIZE(rom_size)) {
 		return rom_size[type];
 	} else if (type >= 0x52 && type < 0x52+ARRAY_SIZE(rom_size_52)) {
